different combinations of string S.txt
In this task, Sarvesh would like for you to appreciate the usefulness of the groupby() function of itertools . 

To read more about this function, Check this out .

You are given a string . 

Suppose a character '' occurs consecutively times in the string. 

Replace these consecutive occurrences of the character '' with in the string.

Constraints:

0≤ S ≤10^4

Input Format:

A single line containing the string S.

Output Format:

Print the different combinations of string S on separate lines.